+++ components/contracts/RequestRevisionModal.tsx
@@ -7,7 +7,6 @@ import React, { useMemo, useState } from "react";
 import {
     Keyboard,
     KeyboardAvoidingView,
-    Modal as RNModal,
     Platform,
     Pressable,
     ScrollView,
@@ -15,11 +14,11 @@ import {
     View,
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import { Modal as PaperModal, Portal } from "react-native-paper";
 import { Text } from "../theme/Themed";
 import Button from "../ui/button";
 import TextInput from "../ui/text-input";
 import { requestVideoRevision } from "./api/review-pending.api";
+import ContractActionOverlay from "./ContractActionOverlay";
 
 export interface RequestRevisionModalProps {
     visible: boolean;
@@ -119,56 +118,19 @@ const RequestRevisionModal: React.FC<RequestRevisionModalProps> = ({
         </KeyboardAvoidingView>
     );
 
-    if (Platform.OS !== "web") {
-        return (
-            <RNModal
-                visible={visible}
-                animationType="slide"
-                onRequestClose={handleClose}
-                statusBarTranslucent
-            >
-                <View style={styles.modalContainer}>{modalContent}</View>
-            </RNModal>
-        );
-    }
-
     return (
-        <Portal>
-            <PaperModal
-                visible={visible}
-                onDismiss={handleClose}
-                contentContainerStyle={styles.modalContainer}
-            >
-                {modalContent}
-            </PaperModal>
-        </Portal>
+        <ContractActionOverlay visible={visible} onClose={handleClose} mode="modal">
+            <View style={styles.modalContainer}>{modalContent}</View>
+        </ContractActionOverlay>
     );
 };
 
 function createStyles(colors: ReturnType<typeof Colors>, safeAreaTop: number) {
-    const isNative = Platform.OS !== "web";
     return StyleSheet.create({
         modalContainer: {
             backgroundColor: colors.background,
-            ...(isNative
-                ? {
-                      flex: 1,
-                      margin: 0,
-                      marginHorizontal: 0,
-                      maxWidth: "100%",
-                      alignSelf: "stretch",
-                      borderRadius: 0,
-                      paddingTop: Math.max(safeAreaTop, 16),
-                      paddingHorizontal: 24,
-                      paddingBottom: 24,
-                  }
-                : {
-                      borderRadius: 12,
-                      padding: 24,
-                      marginHorizontal: 24,
-                      maxWidth: 500,
-                      alignSelf: "center",
-                  }),
+            padding: 24,
+            paddingTop: Math.max(safeAreaTop, 16),
             overflow: "hidden",
         },
         keyboardView: { flex: 1, width: "100%" },
